Switches to fabricatr for simulating data
Closes #30

Switches to using `fabricatr` to simulate data and updates `docs/modelling.qmd` in light of this.

Still have errors in `docs/modelling.qmd` though when attempting to tune the LASSO model the cross-validation fails as
the size of the subset generated differs from the existing data. Confusing as we split data into `train` and `test`
first and set the `recipe` up with just the `train` subset.